move wcaccount to its own class
